ui: move the views init in a dedicated module
Move the UI views initialization and route mapping in a dedicated
function in a new module. This way we can keep `src/index.ts` clean.
